broke ou the webgl code, cleaned it up a bit. also changed the hero text color to black, and added the ability to control the camero on the gems scene
